scripts/gcloud/requirements.txt
requests==2.24.0
beautifulsoup4==4.10.0
google-cloud-storage==1.27.0